Fix pod spec source tag
There are other methods of specifying the location of source code
instead of tags, but you should really be using tags to mark your code
commit for your spec. If you are not the owner of the repository you
are creating a spec for, file an issue with the repository to request
that a tag be made. If this doesn’t work, then you should specify the
version of your podspec to begin at 0.0.1 to indicate that this pod’s
spec is unmanaged. Should the owner finally get a tag, then you can
update the podspec with the correct version with little fear of
conflicting with your previously unmanaged podspec’s version.